Proper support for keystoneauth1 plugins (#4)
* Use project ids instead of project name
* Use Keystone URL instead of nova's
* openstack: load all available authentication plugins
* Load keystoneauth1 session arguments
* Nicer error information when templating
* Make it work with Keystone v2

Avoids having different endpoints for each of the supported
projects. OCCI endpoint still uses the OCCI URL to keep
compatibility.

Instead of only loading the v3password authentication plugin, lets load
all the available keystoneauth plugins so that the user is able to use
their preferred authentication method.

Fixes #86

Small changes to play nicely with Keystone v2 so it works with
Keystone-VOMS authentication that it still the main one in the
production infra.